+++ generator/tools/templates/react/src/App.tsx
@@ -2,9 +2,27 @@ import { useState } from "react";
 import "./App.css";
 import {
     initializeApplet,
+    setOperations
 } from "common/rpc.mjs";
 import useEffectHmr from "./hooks/useEffectOnce";
-import { getLoggedInUser } from "./helpers";
+import { fetchTable, getLoggedInUser } from "./helpers";
+import { Action, Op, Service } from "./contract";
+
+class __contract__Service extends Service {
+
+    @Action
+    increment(num: number) {
+        return { num }
+    }
+
+    @Op()
+    async addNum(num: number) {
+        this.increment(num);
+    }
+
+}
+
+const violetService = new __contract__Service()
 
 
 function App() {
@@ -21,15 +39,39 @@ function App() {
         }
     }
 
+
     useEffectHmr(() => {
-        initializeApplet();
+        initializeApplet(async () => {
+            setOperations(violetService.ops)
+        });
         fetchCurrentUser()
+        fetchTable().then(setCount)
+
     }, []);
 
+
+    const [count, setCount] = useState(0)
+
+
+    const pushAddNumber = async () => {
+        const previousCount = count;
+        await violetService.addNum(num)
+        const newCount = await fetchTable(previousCount)
+        setCount(newCount)
+    }
+
+
+    const [num, setNum] = useState(0)
     return (
         <div className="p-4">
-            <h1>Hello world!</h1>
+            <h1 className="py-4">Hello world!</h1>
             <h2>{error ? error : `Welcome ${currentUser}`}</h2>
+            <div className="p-8">
+
+                <h1>Count: {count}</h1>
+                <input type="number" onChange={(e) => setNum(Number(e.target.value))} />
+                <button className="py-2 rounded-lg text-white bg-blue-500 hover:bg-blue-600 px-4" onClick={() => { pushAddNumber() }}>Increment number!</button>
+            </div>
         </div>
     );
 }
