+++ generator/tools/templates/react/tsconfig.json
@@ -3,6 +3,7 @@
         "target": "ESNext",
         "useDefineForClassFields": true,
         "lib": ["DOM", "DOM.Iterable", "ESNext"],
+        "experimentalDecorators": true,
         "allowJs": false,
         "paths": {
             "/common": ["../../CommonSys/common"]
