--- services/user/Violet/build.sh
@@ -1,4 +1,0 @@
-mkdir build
-cd build
-cmake `psidk-cmake-args` ..
-make -j $(nproc)
