--- services/user/Violet/test/CMakeLists.txt
@@ -1,3 +1,0 @@
-add_user_test_debug(Violet src/Violet-test.cpp)
-
-set(CMAKE_EXPORT_COMPILE_COMMANDS on)
