--- services/user/Violet/test/src/Violet-test.cpp
@@ -1,11 +1,0 @@
-#define CATCH_CONFIG_MAIN
-#include <psio/fracpack.hpp>
-
-#include <psibase/DefaultTestChain.hpp>
-#include <services/system/AccountSys.hpp>
-
-#include "services/user/TokenSys.hpp"
-
-using namespace psibase;
-
-TEST_CASE("Test case 1") {}
