--- services/user/Violet/ui/.prettierrc.json
@@ -1,4 +1,0 @@
-{
-  "tabWidth": 4,
-  "plugins": ["./node_modules/prettier-plugin-tailwindcss"]
-}
