--- services/user/Violet/ui/package.json
@@ -1,36 +1,0 @@
-{
-    "name": "tester",
-    "private": true,
-    "version": "0.0.0",
-    "type": "module",
-    "scripts": {
-        "dev": "vite",
-        "build": "tsc && vite build",
-        "preview": "vite preview"
-    },
-    "dependencies": {
-        "@babel/plugin-proposal-decorators": "^7.19.3",
-        "@rollup/plugin-alias": "^3.1.9",
-        "@types/node": "^18.7.7",
-        "react": "^18.2.0",
-        "react-dom": "^18.2.0",
-        "waait": "^1.0.5"
-    },
-    "devDependencies": {
-        "@tailwindcss/forms": "^0.5.2",
-        "@types/react": "^18.0.17",
-        "@types/react-dom": "^18.0.6",
-        "@vitejs/plugin-react": "^2.0.1",
-        "autoprefixer": "^10.4.8",
-        "eslint": "^8.22.0",
-        "eslint-config-prettier": "^8.5.0",
-        "eslint-plugin-import": "^2.26.0",
-        "postcss": "^8.4.16",
-        "postcss-import": "^14.1.0",
-        "prettier": "^2.7.1",
-        "prettier-plugin-tailwindcss": "^0.1.13",
-        "tailwindcss": "^3.1.8",
-        "typescript": "^4.6.4",
-        "vite": "^3.0.7"
-    }
-}
