--- services/user/Violet/ui/postcss.config.cjs
@@ -1,9 +1,0 @@
-// postcss.config.js
-module.exports = {
-    plugins: {
-        "postcss-import": {},
-        "tailwindcss/nesting": {},
-        tailwindcss: {},
-        autoprefixer: {},
-    },
-};
