--- services/user/Violet/ui/src/main.tsx
@@ -1,8 +1,0 @@
-import ReactDOM from "react-dom/client";
-
-import App from "./App";
-import "./styles/globals.css";
-
-ReactDOM.createRoot(document.getElementById("root") as HTMLElement).render(
-    <App />
-);
