--- services/user/Violet/ui/src/vite-env.d.ts
@@ -1,1 +1,0 @@
-/// <reference types="vite/client" />
