--- services/user/Violet/ui/tsconfig.json
@@ -1,29 +1,0 @@
-{
-    "compilerOptions": {
-        "target": "ESNext",
-        "useDefineForClassFields": true,
-        "lib": ["DOM", "DOM.Iterable", "ESNext"],
-        "experimentalDecorators": true,
-        "allowJs": false,
-        "paths": {
-            "/common": ["../../CommonSys/common"]
-        },
-        "skipLibCheck": true,
-        "esModuleInterop": true,
-        "allowSyntheticDefaultImports": true,
-        "strict": true,
-        "forceConsistentCasingInFileNames": true,
-        "module": "NodeNext",
-        "moduleResolution": "Node",
-        "resolveJsonModule": true,
-        "isolatedModules": true,
-        "noEmit": true,
-        "jsx": "react-jsx"
-    },
-    "include": ["src", "../../CommonSys/common/rpc.d.ts"],
-    "references": [
-        {
-            "path": "./tsconfig.node.json"
-        }
-    ]
-}
