--- services/user/Violet/ui/tsconfig.node.json
@@ -1,9 +1,0 @@
-{
-    "compilerOptions": {
-        "composite": true,
-        "module": "ESNext",
-        "moduleResolution": "Node",
-        "allowSyntheticDefaultImports": true
-    },
-    "include": ["vite.config.ts"]
-}
